Added silver armour having thorns and edited werewolf
Edited some things in EntityWerewolf
Changed onEntityUpdate to onLivingUpdate
Made that silver armour has thorns effect against werewolf
Added logger
Made that every player gets informed when a player has a new texture
Added PureSilverArmorBase
Add IArmorEffectiveAgainstWerewolf
Transformation xp bar now works!
Replaced TwoValueList with HashMap
Did lots of stuff

+++ src/main/java/theblockbox/huntersdream/util/interfaces/IEffectiveAgainstWerewolf.java
@@ -5,10 +5,10 @@ package theblockbox.huntersdream.util.interfaces;
  * against werewolves (= deal more damage)
  */
 public interface IEffectiveAgainstWerewolf {
-	public static final int DEFAULT_EFFECTIVENESS = 2;
+	public static final float DEFAULT_EFFECTIVENESS = 2;
 
 	/** The damage multiplier when used against a werewolf */
-	default public int getEffectiveness() {
+	default public float getEffectiveness() {
 		return DEFAULT_EFFECTIVENESS;
 	}
 }
